Add Diff Output for kmdf_fx2_dmf
The readme previously indicated that a comparison of the original sample (kmdf_fx2) and this sample (kmdf_fx2_dmf) would be useful. While developers certainly are capable doing a directory comparison, having at least the base version available as a part of the documentation is highly useful. This prevents the requirement of syncing a separate repository just to understand this repository, and ensures that the comparison is clear and understandable.